Made changes to the GameMain class. This included initialising the game board and starting the game
Updated the Board class by defining the isDraw method. A nested loop is
used to loop through each cell in the board to check if any cell is
empty.

Defined the constructor public Cell in the Cell class, so that a row and
column indexes are assigned to cells cell.